+++ lib/components/search_homepage.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+
+class SearchHomePage extends StatefulWidget {
+  const SearchHomePage({super.key});
+
+  @override
+  State<SearchHomePage> createState() => _SearchHomePageState();
+}
+
+class _SearchHomePageState extends State<SearchHomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      child: TextFormField(
+        decoration: InputDecoration(
+          filled: true,
+          fillColor: Color.fromARGB(255, 243, 248, 254),
+          prefixIcon: Icon(
+            Icons.search,
+            color: Colors.black.withOpacity(0.3),
+          ),
+          hintText: 'Find things to do',
+          hintStyle: TextStyle(color: Colors.black.withOpacity(0.3)),
+          enabledBorder: OutlineInputBorder(
+            borderRadius: BorderRadius.circular(30),
+            borderSide: BorderSide(
+              color: Color.fromARGB(255, 243, 248, 254),
+            ),
+          ),
+          contentPadding: const EdgeInsets.symmetric(vertical: 15),
+        ),
+      ),
+    );
+  }
+}
